pyproject.toml
[tool.poetry]
name = "fastapi-neon"
version = "0.1.0"
description = ""
authors = ["Mohammad Ali <m.aliadnanrauf@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlmodel = "^0.0.16"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.28.0"}
psycopg = "^3.1.18"
passlib = "^1.7.4"
pyjwt = "^2.8.0"
httpx = "^0.27.0"
pytest = "^8.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
